requirements-dev.txt
-r requirements.txt

pytest==6.0.2
pytest-django==4.1.0
pytest-asyncio==0.14.0
coverage==5.3

bandit==1.6.2
isort==5.6.4
mypy==0.790
safety==1.9.0
cognitive-complexity==1.2.0

mccabe==0.6.1
flake8==3.8.4
flake8-blind-except==0.1.1
flake8-broken-line==0.2.1
flake8-bugbear==20.1.4
flake8-builtins==1.5.3
flake8-class-attributes-order==0.1.1
flake8-cognitive-complexity==0.1.0
flake8-commas==2.0.0
flake8-comprehensions==3.2.3
flake8-debugger==3.2.1
flake8-eradicate==1.0.0
flake8-functions==0.0.4
flake8-isort==4.0.0
flake8-mock==0.3
flake8-mutable==1.2.0
flake8-print==3.1.4
flake8-pytest==1.3
flake8-pytest-style==1.3.0
flake8-quotes==3.2.0
flake8-string-format==0.3.0
flake8-variables-names==0.0.3
